--- web/src/components/HelpBox.test.js
@@ -1,15 +1,0 @@
-import { shallow } from 'enzyme';
-import React from 'react';
-
-import HelpBox from './HelpBox';
-
-describe('HelpBox component', () => {
-  test('renders correctly', () => {
-    const component = shallow(
-      <HelpBox>
-        <div className="test child" />
-      </HelpBox>
-    );
-    expect(component).toMatchSnapshot();
-  });
-});
