--- web/src/components/StayTuned.js
@@ -1,10 +1,0 @@
-import React from 'react';
-
-import Md from './Md';
-import { t } from '../i18n';
-
-const StayTuned = () => (
-  <Md content={t('wip')} wrapper="div" className="p2 alert alert-info" />
-);
-
-export default StayTuned;
